+++ src/invoices/InvoiceDetails/utils.js
@@ -70,6 +70,12 @@ export const showUpdateInvoiceError = async (
       showCallout({ messageId: `ui-invoice.invoice.actions.approve.error.${ERROR_CODES[code]}`, type: 'error', values: { fundCodes } });
       break;
     }
+    case ERROR_CODES.incorrectFundDistributionTotal: {
+      const invoiceLineNumber = error?.errors?.[0]?.parameters?.filter(({ key }) => key === 'invoiceLineNumber')[0]?.value;
+
+      showCallout({ messageId: `ui-invoice.invoice.actions.approve.error.${ERROR_CODES[code]}`, type: 'error', values: { invoiceLineNumber } });
+      break;
+    }
     case ERROR_CODES.budgetExpenseClassNotFound: {
       const fundCode = error?.errors?.[0]?.parameters?.filter(({ key }) => key === 'fundCode')[0]?.value;
       const expenseClassName = error?.errors[0]?.parameters?.filter(({ key }) => key === 'expenseClassName')[0]?.value;